requirements.txt
Flask==3.1.0
sqlalchemy==2.0.38
python-dotenv==1.0.1
requests==2.32.3
openai==1.65.2

celery
redis
psycopg2-binary
pydantic
